users added to groups, groups can be deleted

+++ Backend/server/src/main/java/coms309/directmessagegroup/DirectMessageGroupController.java
@@ -1,6 +1,7 @@
-package coms309.group;
+package coms309.directmessagegroup;
 
 import java.util.List;
+import java.util.Set;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,40 +25,59 @@ import io.swagger.annotations.ApiOperation;
 @Api(value = "DirectMessageController", description = "REST APIs for directmessage entity")
 @RequestMapping("/api/groups")
 @RestController
-public class GroupController {
+public class DirectMessageGroupController {
     @Autowired
-    GroupService groupService;
+    DirectMessageGroupService groupService;
 
     @Autowired
     UserService userService;
 
-    Logger logger = LoggerFactory.getLogger(GroupController.class);
+    Logger logger = LoggerFactory.getLogger(DirectMessageGroupController.class);
 
     @ApiOperation(value = "lists all groups", response = Iterable.class, tags = "getAllGroups")
     @GetMapping
-    private List<Group> getAllGroups() {
+    private List<DirectMessageGroup> getAllGroups() {
         logger.info("got all groups");
         return groupService.getAllGroups();
     }
 
-    @ApiOperation(value = "gets a group by id", response = Group.class, tags = "getGroup")
+    @ApiOperation(value = "gets a group by id", response = DirectMessageGroup.class, tags = "getGroup")
     @GetMapping("/{groupId}")
-    private Group getGroup(@PathVariable("groupId") int id) {
-        logger.info("got group by id");
-        return groupService.getGroupById(id);
+    private DirectMessageGroup getGroup(@PathVariable("groupId") int id) {
+        DirectMessageGroup group = groupService.getGroupById(id);
+        if(group != null) {
+            logger.info("got group by id");
+            return groupService.getGroupById(id);
+        }
+
+        return null;
     }
 
     @ApiOperation(value = "delete a group by id", response = String.class, tags = "deleteGroup")
-    @DeleteMapping
+    @DeleteMapping("/{groupId}")
     private String deleteGroup(@PathVariable("groupId") int id) {
         logger.info("deleting group");
-        groupService.delete(id);
-        return "deleted group: " + id;
+        DirectMessageGroup group = groupService.getGroupById(id);
+        if(group != null) {
+            Set<User> usersInGroup = group.getMembers();
+
+            for (User user : usersInGroup) {
+                user.removeFromGroup(group);
+                userService.saveOrUpdate(user);
+            }
+
+            group.dismissMembers();
+            groupService.delete(id);
+            return "deleted group: " + id;
+        }
+
+        return "group not found";
+
     }
 
-    @ApiOperation(value = "creates a new group", response = Group.class, tags = "saveGroup")
+    @ApiOperation(value = "creates a new group", response = DirectMessageGroup.class, tags = "saveGroup")
     @PostMapping
-    private Group saveGroup(@RequestBody Group group) {
+    private DirectMessageGroup saveGroup(@RequestBody DirectMessageGroup group) {
         logger.info("created a new group");
         groupService.saveOrUpdate(group);
         return group;
@@ -66,7 +86,7 @@ public class GroupController {
     @ApiOperation(value = "adds a member to a group", response = String.class, tags = "addMemberToGroup")
     @PatchMapping("/{groupId}/add/{userId}")
     private String addMemberToGroup(@PathVariable("groupId") int id, @PathVariable("userId") int userId) {
-        Group group = groupService.getGroupById(id);
+        DirectMessageGroup group = groupService.getGroupById(id);
         User user = userService.getUserById(userId);
 
         if(group == null) {
@@ -78,6 +98,7 @@ public class GroupController {
         }
 
         group.addMemberToGroup(user);
+        user.addGroup(group);
         groupService.saveOrUpdate(group);
         return "User " + user.getId() + " added to group " + group.getName();
     }
@@ -85,7 +106,7 @@ public class GroupController {
     @ApiOperation(value = "removes a member from a group", response = String.class, tags = "removeMemberFromGroup")
     @PatchMapping("/{groupId}/remove/{userId}")
     private String removeMemberFromGroup(@PathVariable("groupId") int id, @PathVariable("userId") int userId) {
-        Group group = groupService.getGroupById(id);
+        DirectMessageGroup group = groupService.getGroupById(id);
         User user = userService.getUserById(userId);
 
         if(group == null) {
